Add generateRequests_Dynamic_Traffic & Hold_time
Added hold_time variable to Request class. Variable is the duration serviced connection is held in the game.
set the DEFAULT value to 10 in generateRequests() function as per original.
&
Added function which allows specific selection of request interval and hold time for the requests to allow for game to be played with dynamic traffic loads.